insert-new-data-endpoint/insert-jobs-endpoint/requirements.txt
functions-framework
#flask
python-tds
sqlalchemy-pytds
SQLAlchemy
cloud-sql-python-connector